requirements.txt
alive_progress==3.1.5
bitstring==4.2.2
crc==7.0.0
debugpy==1.8.1
matplotlib==3.8.3
pyserial==3.5
PySide2==5.15.2.1
PyVISA==1.14.1
PyYAML==6.0.2
scipy==1.14.1
